tests: Move common deposit operations to their own module.
