+++ core/tests/common/mod.rs
@@ -1,0 +1,5 @@
+//! # Common Utilities for Integration Tests
+
+mod deposit;
+
+pub use deposit::*;
